backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
beautifulsoup4
pandas
scikit-learn
xgboost
numpy
joblib
python-dotenv
psycopg2-binary
